+++ projects/ngz-material/src/lib/components/grid/index.ts
@@ -4,6 +4,7 @@
 // Import modules
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { RouterModule } from '@angular/router';
 import { MatTableModule } from '@angular/material/table';
 import { MatSortModule } from '@angular/material/sort';
 import { MatPaginatorModule } from '@angular/material/paginator';
@@ -14,6 +15,7 @@ import { MatInputModule } from '@angular/material/input';
 import { MatTooltipModule } from '@angular/material/tooltip';
 const modules = [
   CommonModule,
+  RouterModule,
   MatTableModule,
   MatSortModule,
   MatPaginatorModule,